注解限流生效，不能被类内部调用
